+++ packages/ag-grid-enterprise/src/rowHierarchy/rowHierarchyModule.ts
@@ -7,8 +7,8 @@ import { RowGroupColsSvc } from '../rowGrouping/rowGroupColsSvc';
 import { VERSION } from '../version';
 import { AutoColService } from './autoColService';
 import { ClientSideExpansionService } from './clientSideExpansionService';
-import { GroupHierarchyColService } from './dateHierarchyColService';
 import { FlattenStage } from './flattenStage';
+import { GroupHierarchyColService } from './groupHierarchyColService';
 import { GroupStage } from './groupStage';
 import { GroupCellRenderer } from './rendering/groupCellRenderer';
 import { GroupCellRendererCtrl } from './rendering/groupCellRendererCtrl';
